+++ IcVibracoes.Core/NumericalIntegrationMethods/NewmarkBeta/NewmarkBetaMethod.cs
@@ -1,6 +1,6 @@
 ﻿using IcVibracoes.Core.Calculator.Force;
 using IcVibracoes.Core.DTO;
-using IcVibracoes.Core.DTO.NumericalMethodInput.FiniteElement;
+using IcVibracoes.Core.DTO.NumericalMethodInput.FiniteElements;
 using IcVibracoes.Core.DTO.NumericalMethodInput.RigidBody;
 using IcVibracoes.Core.ExtensionMethods;
 using IcVibracoes.Core.Models;