+++ IcVibracoes.Core/NumericalIntegrationMethods/RungeKuttaForthOrder/RungeKuttaForthOrderMethod.cs
@@ -1,6 +1,6 @@
 ﻿using IcVibracoes.Core.Calculator.DifferentialEquationOfMotion;
 using IcVibracoes.Core.DTO;
-using IcVibracoes.Core.DTO.NumericalMethodInput.FiniteElement;
+using IcVibracoes.Core.DTO.NumericalMethodInput.FiniteElements;
 using IcVibracoes.Core.DTO.NumericalMethodInput.RigidBody;
 using System;
 using System.Threading.Tasks;