+++ IcVibracoes.Core/Operations/CalculateVibration/FiniteElement/ICalculateVibration_FiniteElements.cs
@@ -1,5 +1,5 @@
 ﻿using IcVibracoes.Common.Profiles;
-using IcVibracoes.Core.DTO.NumericalMethodInput.FiniteElement;
+using IcVibracoes.Core.DTO.NumericalMethodInput.FiniteElements;
 using IcVibracoes.Core.Models.Beams;
 using IcVibracoes.DataContracts.FiniteElement;
 using System.Threading.Tasks;